[#]改变文件结构 引用方式 由require 改为import，更符合使用习惯 [#]定义参数并声明属性 [#]增加参数 bannerAction, 触发头部点击事件的回调
